Add a Noop Parser
This Noop is used when running with `-loop 0` when you want to run the REST server only.
It avoids a NPE when checking if the parser is still running or not.

Closes #608.
